car_fleet_service/requirements.txt
django
djangorestframework
gunicorn
requests